Move HttpClient file to path supported by the autoloader
